Fix all build issues warning in SDK
Squashed commit of the following:

commit b93393eb8f8320fb2791f7f8da36f69c01a17307
Author: Abhishek Sharma <asharma@appnexus.com>
Date:   Wed Nov 6 21:08:46 2019 +0530

    Updated OMID Lib header

commit a6408d7f8121249a491b3b58ca47e2b767767a4f
Merge: 7e1a5ba f1117ad
Author: Abhishek Sharma <asharma@appnexus.com>
Date:   Wed Nov 6 21:05:55 2019 +0530

    Merged develop

commit 7e1a5ba32f7ea5640c7956ad46886ae56ad5fe04
Author: Abhishek Sharma <asharma@appnexus.com>
Date:   Fri Oct 4 14:10:45 2019 +0530

    Fixed Cast issue

commit 8d2149cb51f8db48262b1f7ea5392f730b7931ce
Merge: 7e6474d bb4c27e
Author: Abhishek Sharma <asharma@appnexus.com>
Date:   Fri Oct 4 13:24:17 2019 +0530

    Merge remote-tracking branch 'origin/develop' into MS_3997_Fix_all_build_issues_warning_in_SDK

commit 7e6474ddd4a948c518af02bd8bf7ffa8c8f1a327
Author: Abhishek Sharma <asharma@appnexus.com>
Date:   Thu Oct 3 19:06:45 2019 +0530

    Updated Logic

commit dbcd968bd525f97a1f959017a9d2850bd72be209
Author: Abhishek Sharma <asharma@appnexus.com>
Date:   Thu Oct 3 16:59:20 2019 +0530

    Updated openURL Logic

commit c216925bcc1533b061fb5e98853028569001c324
Author: Abhishek Sharma <asharma@appnexus.com>
Date:   Fri Sep 13 17:08:21 2019 +0530

    Updated Native mediation init

commit 2b5208c51ea9d70685f6d5588c382d647bf972bb
Author: Abhishek Sharma <asharma@appnexus.com>
Date:   Thu Sep 12 16:05:23 2019 +0530

    Updated customKeywords

commit 0e3bcd65962154219738897bd78735ab8e77a8ca
Author: Abhishek Sharma <asharma@appnexus.com>
Date:   Thu Sep 12 14:55:18 2019 +0530

    Fixed Warning

commit 14144ecdc56fd11e1d0d5a4b7ebd4d9ca5d1ca3e
Author: Abhishek Sharma <asharma@appnexus.com>
Date:   Tue Sep 10 20:18:51 2019 +0530

    Fixed Build Error

+++ sdk/sourcefiles/internal/ANMediationAdViewController.m
@@ -371,7 +371,7 @@
     [self runInBlock:^(void) {
         ANUniversalAdFetcher *fetcher = self.adFetcher;
         
-        NSString *responseURL = [self.mediatedAd.responseURL an_responseTrackerReasonCode:errorCode
+        NSString *responseURL = [self.mediatedAd.responseURL an_responseTrackerReasonCode:(int)errorCode
                                                                                   latency: ([self getLatency] * 1000)
                                                                              totalLatency:([self getTotalLatency] * 1000)];
